add support for choosing the concatenation dim when computing affs.
We want to be able to process a set of affs on data with an optional batch dimension, and an optional channel dimension. Thus depending on the semantics of the first dimension, we often want to concatenate in either the 0 or 1 dimension. But we can't know that purely from the shape of the array. Thus it can be user specified